Move: 이미지 서비스 클래스 구조 변경에 따른 import 패키지 변경
